refactor(invoices): extract outcomes lifecycle slice
