+++ backend/app/features/invoices/outcomes/__init__.py
@@ -1,0 +1,5 @@
+"""Invoice outcome lifecycle slice."""
+
+from app.features.invoices.outcomes.service import InvoiceOutcomeService
+
+__all__ = ["InvoiceOutcomeService"]
